requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
pylint==3.0.3
autopep8==2.0.4
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic==2.6.1
ast-comments==1.0.1
black==24.2.0
pygments>=2.17.0
aiohttp>=3.9.0 